JIT buffer relocation and 2% PHP8.1.4 performance
This is a JIT buffer relocation experiment inspired by this blog
https://v8.dev/blog/short-builtin-calls
  For 64-bit applications, branch prediction performance can be
  negatively impacted when the target of a branch is more than
  4 GB away from the branch.

We set up a big 500MB char array which is close to PHP .text segment
and use it as our own JIT buffer (originally allocated by opcache.so).

In our benchmark, we found PHP interpreter archieved 2% performance
and much better branching performance.

Signed-off-by: Tony <tao.su@intel.com>
Signed-off-by: Xue Wang <xue1.wang@intel.com>